backend/requirements.txt
Django>=5.1,<5.2
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.4
django-csp>=3.8
django-environ>=0.11
whitenoise>=6.7
requests>=2.32
gunicorn>=22.0
